+++ lib/creator/transformations/index.js
@@ -1,0 +1,25 @@
+const chalk = require('chalk');
+const validateSchema = require('../../utils/validateSchema.js');
+const webpackOptionsSchema = require('../../utils/webpackOptionsSchema.json');
+const WebpackOptionsValidationError = require('../../utils/WebpackOptionsValidationError');
+/*
+* @function runTransform
+*
+* Runs the transformations from an object we get from yeoman
+*
+* @param { Object } transformObject - Options to transform
+* @returns { <Void> } TODO
+*/
+
+module.exports = function runTransform(transformObject) {
+
+		// scaffold is done
+		/*
+		const webpackOptionsValidationErrors = validateSchema(webpackOptionsSchema, initialWebpackConfig);
+		if (webpackOptionsValidationErrors.length) {
+			throw new WebpackOptionsValidationError(webpackOptionsValidationErrors);
+		} else {
+			process.stdout.write('\n' + chalk.green('Congratulations! Your new webpack config file is created!') + '\n');
+		}
+		*/
+};
