+++ lib/creator/utils/validate-options.js
