+++ lib/creator/utils/validate-options.spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const validateOptions = require('../../__mocks__/creator/validate-options.mock').validateOptions;
+const validateOptions = require('../../../__mocks__/creator/validate-options.mock').validateOptions;
 
 describe('validate-options', () => {
