+++ lib/creator/yeoman/webpack-adapter.js
