+++ lib/creator/yeoman/webpack-generator.js
