+++ lib/initialize.js
@@ -8,8 +8,8 @@ const creator = require('./creator/index');
 * if we are running the init command with no arguments or if we got dependencies
 *
 * @param { Object } pkg - packages included when running the init command
-* @returns { <Function> } prompt|npmPackagesExists - returns either an inquirer prompt with
-* the initial questions we provide, or validates the packages given
+* @returns { <Function> } creator|npmPackagesExists - returns an installation of the package,
+* followed up with a yeoman instance of that if there's packages. If not, it creates a defaultGenerator
 */
 
 module.exports = function initializeInquirer(pkg) {
