+++ lib/migrate.js
@@ -4,6 +4,9 @@ const diff = require('diff');
 const inquirer = require('inquirer');
 const PLazy = require('p-lazy');
 const Listr = require('listr');
+const validateSchema = require('./utils/validateSchema.js');
+const webpackOptionsSchema = require('./utils/webpackOptionsSchema.json');
+const WebpackOptionsValidationError = require('./utils/WebpackOptionsValidationError');
 
 module.exports = function transformFile(currentConfigPath, outputConfigPath, options) {
 	const recastOptions = Object.assign({
@@ -65,9 +68,17 @@ module.exports = function transformFile(currentConfigPath, outputConfigPath, opt
 				])
 				.then(answers => {
 					if (answers['confirmMigration']) {
-						// TODO validate the config
-						fs.writeFileSync(outputConfigPath, result, 'utf8');
-						console.log(chalk.green(`✔︎ New webpack v2 config file is at ${outputConfigPath}`));
+						fs.writeFile(outputConfigPath, result, 'utf8', (err) => {
+							const webpackOptionsValidationErrors = validateSchema(webpackOptionsSchema, require(outputConfigPath));
+							if (err) {
+								throw err;
+							}
+							else if (webpackOptionsValidationErrors.length) {
+								throw new WebpackOptionsValidationError(webpackOptionsValidationErrors);
+							} else {
+								console.log(chalk.green(`\n ✔︎ New webpack v2 config file is at ${outputConfigPath}`));
+							}
+						});
 					} else {
 						console.log(chalk.red('✖ Migration aborted'));
 					}
