+++ lib/utils/WebpackOptionsValidationError.js
