+++ lib/utils/resolve-packages.js
@@ -49,7 +49,7 @@ module.exports = function resolvePackages(pkg) {
 	return processPromise(spawnChild(pkg)).then( () => {
 		try {
 			let loc = path.join('..', '..', 'node_modules', pkg);
-			creator(loc, null);
+			return creator(loc);
 		} catch(err) {
 			console.log('Package wasn\'t validated correctly..');
 			console.log('Submit an issue for', pkg, 'if this persists');
