+++ lib/utils/validateSchema.js
