+++ lib/utils/webpackOptionsSchema.json
